+++ instance/emf/tutorial/level/smallintasbooleantype/src/main/resources/data/Fact.csv
